requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
dj-database-url==3.1.1
Django==5.2.10
django-cors-headers==4.3.1
django-unfold==0.80.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.11
packaging==26.0
pillow==10.2.0
psycopg2-binary==2.9.9
PyJWT==2.11.0
python-decouple==3.8
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.6.0
